Delete the class AITest. Move my test to the class JavAngry. Add extends to JavAngryTest
