+++ src/test/java/school/redrover/JavAngryTest.java
@@ -7,13 +7,14 @@ import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.interactions.Actions;
 import org.testng.Assert;
 import org.testng.annotations.Test;
+import school.redrover.runner.BaseTest;
 
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-public class JavAngryTest {
+public class JavAngryTest extends BaseTest {
 
     @Test
     void testPriceCheckMainPageShoppingPage() {
@@ -189,6 +190,19 @@ public class JavAngryTest {
         driver.findElement(By.xpath("//input[@id='payment_method_cod']")).click();
         driver.findElement(By.xpath("//button[@name='woocommerce_checkout_place_order']")).click();
         driver.quit();
+    }
+
+    @Test
+    void testFields() {
+        WebDriver webDriver = getDriver();
+        webDriver.get("https://letcode.in/test");
+
+        webDriver.findElement(By.xpath("//a[contains(text(),'Edit')]")).click();
+        webDriver.findElement(By.id("fullName")).sendKeys("Ali Ian");
+        webDriver.findElement(By.id("join")).sendKeys(" and happy \t");
+        webDriver.findElement(By.id("clearMe")).clear();
+
+        Assert.assertTrue(webDriver.findElement(By.id("noEdit")).isDisplayed());
     }
 }
